统一ExamLab项目中Number类型为double类型以支持小数点运算
- 修改OperationPointEditDialogViewModel中的Number类型验证逻辑，从int.TryParse改为double.TryParse
- 修改PowerPointModuleViewModel中的Number类型验证逻辑，从int.TryParse改为double.TryParse
- 修改WindowsModuleViewModel中的Number类型验证逻辑，从int.TryParse改为double.TryParse
- 修改OperationPointEditPage.xaml.cs中CreateNumberControl方法，从int.TryParse改为double.TryParse
- 统一使用Math.Abs进行-1通配符值的浮点数比较
- 确保所有Number类型参数处理逻辑一致，支持小数点数值运算
- ConfigurationParameter类的MinValue和MaxValue属性已为double类型
- ValidationService和其他ViewModel已正确使用double类型

+++ ExamLab/Views/OperationPointEditPage.xaml.cs
@@ -313,11 +313,11 @@ public sealed partial class OperationPointEditPage : Page
             HorizontalAlignment = HorizontalAlignment.Stretch
         };
 
-        if (int.TryParse(parameter.Value, out int currentValue))
+        if (double.TryParse(parameter.Value, out double currentValue))
         {
             numberBox.Value = currentValue;
         }
-        else if (int.TryParse(parameter.DefaultValue, out int defaultValue))
+        else if (double.TryParse(parameter.DefaultValue, out double defaultValue))
         {
             numberBox.Value = defaultValue;
         }
